refactor: Merge friend module into identity
